Add files via upload
This is a project on data modeling with Postgres. I have build an ETL pipeline using Python. The project will define fact and dimension tables for a star schema for a particular analytic focus, and an ETL pipeline that transfers data from files in two local directories into the tables in Postgres using Python and SQL.